+++ NEWS.md
@@ -4,9 +4,9 @@
 * Added classes `BaseSets`.
 * Added `relations`, `elementData`, and `setData` accessors for class `BaseSets`.
 * Added `length`, `nElements`, `nSets`, `setLengths` and `elementLengths` methods for class `BaseSets`.
-* Added method `subset` for classes `BaseSets`.
+* Added method `subset` for class `BaseSets`.
 * Added classes `FuzzyHits` and `FuzzySets`.
-* Added `membership` accessors for class `FuzzySets`.
+* Added `membership` accessor for class `FuzzySets`.
 * Added classes `GOHits` and `GOSets`.
 * Added `evidence` and `ontology` accessors for class `GOSets`.
 * Added `import` and `export` methods for GMT file format.